+++ server/src/models/diesel.rs
@@ -5,7 +5,7 @@ use serde::{Deserialize, Serialize};
 pub(crate) struct MinaProposal {
     pub(crate) id: i32,
     pub(crate) key: String,
-    pub(crate) global_start_slot: i32,
-    pub(crate) global_end_slot: i32,
+    pub(crate) global_start_slot: i64,
+    pub(crate) global_end_slot: i64,
     pub(crate) ledger_hash: Option<String>,
 }
